added global constants and interfaces
