+++ FE/angular/src/app/constants/index.ts
@@ -1,0 +1,5 @@
+export const API_URL_CONST = {
+  posterUrl: 'https://m.media-amazon.com/images/M/',
+  replacePosterUrl: '/assets/images/',
+  serviceUrl: 'https://www.omdbapi.com/?apikey=f59b2e4b&'
+};
