+++ FE/angular/src/styles/main.scss
@@ -1,4 +1,3 @@
 @import 'abstracts/variables';
 @import 'layout/elements';
 @import 'components/movie';
-
